多模态/agents-langgraph-chrome/requirements.txt
langgraph==0.2.36
langchain==0.2.14
openai==1.43.0
playwright==1.46.0
paddlepaddle==2.6.1; platform_system != 'Darwin' or platform_machine != 'arm64'
paddleocr==2.7.3
Pillow==10.4.0
numpy==1.26.4
anyio==4.4.0
rich==13.7.1
typer==0.12.3

